Migrate `Double` to `BigDecimal` for improved precision in financial calculations across models and strategies
- Replace `Double` with `BigDecimal` in key data classes such as `Position`, `AllocationRowViewModel`, and `PositionRecommendation`.
- Update all related calculations, data bindings, and serializers for compatibility with `BigDecimal`.
- Introduce `KBigDecimal` type alias and custom serializer for seamless JSON serialization/deserialization.
- Refactor mappers, strategy logic, and UI components to handle `BigDecimal` values accurately.

+++ src/main/kotlin/com/droidkfx/st/position/AccountPosition.kt
@@ -2,13 +2,14 @@ package com.droidkfx.st.position
 
 import com.droidkfx.st.account.Account
 import com.droidkfx.st.strategy.PositionRecommendation
+import java.math.BigDecimal
 
 data class AccountPosition(
     val account: Account,
     val positionTargets: List<PositionTarget>,
     val currentPositions: List<Position>,
     val currentRecommendedChanges: List<PositionRecommendation>,
-    val currentCash: Double,
+    val currentCash: BigDecimal,
 )
 
 fun AccountPosition.withNewPositionTargets(newPositionTargets: List<PositionTarget>) =
